feat: implement useAutoComplete composable for enhanced autocomplete functionality
